+++ simulariumio/cytosim/cytosim_converter.py
@@ -7,7 +7,7 @@ import sys
 
 import numpy as np
 
-from ..custom_converter import TrajectoryConverter
+from ..trajectory_converter import TrajectoryConverter
 from ..data_objects import TrajectoryData, AgentData, UnitData
 from ..exceptions import DataError
 from ..constants import VIZ_TYPE
