+++ simulariumio/data_objects/__init__.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from .agent_data import AgentData  # noqa: F401
-from .custom_data import TrajectoryData  # noqa: F401
+from .trajectory_data import TrajectoryData  # noqa: F401
 from .unit_data import UnitData  # noqa: F401
 from .histogram_plot_data import HistogramPlotData  # noqa: F401
 from .scatter_plot_data import ScatterPlotData  # noqa: F401
