+++ simulariumio/tests/converters/test_custom_converter.py
@@ -1095,6 +1095,6 @@ from simulariumio.tests.conftest import three_default_agents
 )
 def test_custom_trajectory_reader(trajectory, expected_data):
     converter = TrajectoryConverter(trajectory)
-    buffer_data = converter._read_custom_data(converter._data)
+    buffer_data = converter._read_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
