+++ simulariumio/tests/converters/test_readdy_converter.py
@@ -204,6 +204,6 @@ from simulariumio import UnitData
 )
 def test_cytosim_trajectory_reader(trajectory, expected_data):
     converter = ReaddyConverter(trajectory)
-    buffer_data = converter._read_custom_data(converter._data)
+    buffer_data = converter._read_trajectory_data(converter._data)
     assert expected_data == buffer_data
     assert converter._check_agent_ids_are_unique_per_frame(buffer_data)
